refactor(test): move all integration tests in a 'integration' folder
